Refactor the patterns
